requirements.txt
streamlit 
google-generativeai 
python-dotenv 
pdf2image 
docx
fpdf
pillow
python-docx
fpdf
requests